Test handling of responses that contain invalid data
