+++ spec/models/asteroid/near_earth_object_spec.rb
@@ -7,6 +7,7 @@ describe Asteroid::NearEarthObject do
 
   describe 'validations' do
     it { is_expected.to validate_presence_of :name }
+    it { is_expected.to validate_presence_of :neo_reference_id }
   end
 
   describe 'associations' do
