+++ spec/support/helpers/api_responses/neo/twenty_valid.txt
